@robodoo 2024-12-04
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/189576.png)](https://mergebot.odoo.com/odoo/odoo/pull/189576)